requirements.txt
dnspython>=2.0.0
colorama>=0.4.6
requests>=2.31.0